requirements.txt
openai>=1.40.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyarrow>=16.0.0
